==> Drug_Repositioning/6_new_disease/1_OMIM_disease_parsing_(stop_phenotypic serires).ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import requests
import time

asso = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/bi_random_walk_Cdataset_association.tsv")

# read the mimTitles file
mimTitles = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/6_new_disease/mimTitles.txt", skiprows = 2)
mimTitles = mimTitles.rename(columns = {"# Prefix" : "Prefix","Mim Number" : "Mim_Number", "Preferred Title; symbol" : "Preferred_Title"})
print len(mimTitles)
mimTitles.head(1)

# read the morbidmap file
morbid = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/6_new_disease/morbidmap_txt.txt")
print len(morbid)
morbid.head(1)

print len(asso)
asso.head(1)

di_mimNum = asso['disease_id'].drop_duplicates().str[1:]

int_di_mimNum = list()
for i in di_mimNum:
    int_di_mimNum.append (int(i))

df_disease = mimTitles[mimTitles['Mim_Number'].isin(int_di_mimNum)]

print len(df_disease)
df_disease.head(1)

pair= pd.DataFrame()
for index, row in df_disease.iterrows():
    mim_number = row['Mim_Number']
    title = row['Preferred_Title']
    temp_df = pd.DataFrame([[mim_number, title]], columns = ['Mim_Number','Preferred_Title'])
    pair = pair.append(temp_df)

print len(pair)
pair.head(1)

di_phen = pd.DataFrame()
result = pd.DataFrame()
temp = pd.DataFrame()
cnt = 0
for index, row in pair.iterrows():
    di_mim_number = row['Mim_Number']
    found = morbid[morbid['Phenotype'].str.contains(str(di_mim_number))]
    if found.empty :
        continue
    else:
        cnt = cnt + 1
        phe_mim_number = found['MIM_Number']
        #temp_df = pd.DataFrame(data = [[di_mim_number, phe_mim_number]], columns = ['disease', 'phenotype'])
        temp_df = pd.DataFrame({'disease' : di_mim_number, 'phenotype' : phe_mim_number})
        temp_df2 = pd.DataFrame({'disease' : di_mim_number, 'disease_name' : row['Preferred_Title'], 'phenotype' : phe_mim_number, 'phenotype_title' : found['Phenotype']})
        di_phen = di_phen.append(temp_df)
        result = result.append(found)
        temp = temp.append(temp_df2)

temp.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/test.tsv", sep='\t', index=False)

cnt

print len(result)
result.head()

print len(di_phen)
di_phen.head()

print len(di_phen.drop_duplicates(subset = 'disease'))
print len(di_phen.drop_duplicates(subset = 'phenotype'))

left = pair[~pair['Mim_Number'].isin(di_phen['disease'].drop_duplicates())]

left.to_csv("/home/share/aschoi/data/Drug_Repositioning/6_new_disease/temp2.tsv", sep="\t", index=False)

result1 = result

cnt = 0 
result2 = pd.DataFrame()
temp = pd.DataFrame()
for index, row in left.iterrows():
    di_mim_number = row['Mim_Number']
    title = row['Preferred_Title'].split(', ')[0]
    found = morbid[morbid['Phenotype'].str.contains(title, case=False)]
    if found.empty :
        continue
    else:
        cnt = cnt + 1
        phe_mim_number = found['MIM_Number']
        #temp_df = pd.DataFrame(data = [[di_mim_number, phe_mim_number]], columns = ['disease', 'phenotype'])
        #temp_df = pd.DataFrame({'disease' : di_mim_number, 'phenotype' : phe_mim_number})
        temp_df = pd.DataFrame({'disease' : di_mim_number, 'disease_name' : row['Preferred_Title'], 'phenotype' : phe_mim_number, 'phenotype_title' : found['Phenotype']})
        #di_phen = di_phen.append(temp_df)
        temp = temp.append(temp_df)
        result2 = result2.append(found)

title

cnt

print len(result2)
result2.head()

temp.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/test.tsv", sep='\t', index=False)

print len(temp)
temp

found = morbid[morbid['Phenotype'].str.contains('ALOPECIA', case=False)]
found

len(found)
