==> Drug_Repositioning/6_new_disease/2_OMIM_UMLS_matching.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import requests
import time

mapping = pd.read_excel("/home/share/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_S6 Mapping between OMIM disease and UMLS concepts.xls")

print len(mapping)
mapping.head()

disease = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/4_bi_random_walk_Cdataset_association.tsv")

print len(disease)
disease.head(1)

di_mimNum = disease['disease_id'].drop_duplicates().str[1:]
int_di_mimNum = list()
for i in di_mimNum:
    int_di_mimNum.append (int(i))
print len(int_di_mimNum)

omim_umls = mapping[mapping['OMIM ID'].isin(int_di_mimNum)]
print len(omim_umls)
omim_umls.head()

omim_umls.to_csv("/home/share/aschoi/data/Drug_Repositioning/6_new_disease/omim_umls_mapping_association.tsv", sep='\t', index=False)
omim_umls.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/omim_umls_mapping_association.tsv", sep='\t', index=False)
