==> Drug_Repositioning/6_new_disease/4_DisGeNet.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import requests
import time

asso = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/4_bi_random_walk_Cdataset_association.tsv")

omim_umls = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/6_new_disease/omim_umls_mapping_association.tsv")

disgen = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/6_new_disease/curated_gene_disease_associations.tsv", skiprows = 21)

disgen.head(1)

disgen[disgen['diseaseId'].str.contains('C0035258')].head(1)

omim_umls.head(1)

result = pd.DataFrame()
for index, row in omim_umls.iterrows():
    umls = row['UMLS concept ID']
    found = disgen[disgen['diseaseId'].str.contains(umls)]
    if found.empty :
        continue
    else:
        for index2, row2 in found.iterrows():
            temp_df = pd.DataFrame(data = [[row['OMIM ID'], row['OMIM disease name'], umls, row['UMLS concept name'], row2['geneId'], row2['score'], row2['geneName'], row2['sourceId']]], 
                                       columns = ['OMIM_ID', 'OMIM_disease_name', 'UMLS_concept_ID', 'UMLS_concept_name', 'DisGeNet_geneId', 'DIsGeNet_score', 'DIsGeNet_geneName', 'DIsGeNet_sourceId'])
            result = result.append(temp_df) 

len(result['OMIM_ID'].drop_duplicates())

result.head(1)

association 364개 중에 257개 가 얼마나 해당하는지 찾아보기

filterd_asso = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/5_filtered_association(drug).tsv")

print len(filterd_asso)
filterd_asso.head(1)

di_mimNum = filterd_asso['disease_id'].drop_duplicates().str[1:]
int_di_mimNum = list()
for i in di_mimNum:
    int_di_mimNum.append (int(i))
print len(int_di_mimNum)
filtered = result[result['OMIM_ID'].isin(int_di_mimNum)]
print len(filtered)
print len(filtered.drop_duplicates(subset = 'OMIM_ID'))
filtered.head()

len(filtered['DisGeNet_geneId'].drop_duplicates())

235개가 해당

그럼 이 235개와 연관이 있는 drug의 수와 association의 수는?! --> 이게 gold standard 최종이 될것이다. 

gold = filterd_asso[filterd_asso['disease_id'].isin(filtered['OMIM_ID'])]

print len(gold)
print len(gold['disease_id'].drop_duplicates())
print len(gold['drug_id'].drop_duplicates())

gold.to_csv("/home/share/aschoi/data/Drug_Repositioning/7_new_association/7_gold_association.tsv", sep='\t', index=False)
gold.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/7_gold_association.tsv", sep='\t', index=False)

gold.head()

filtered.to_csv("/home/share/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_disease_feature.tsv", sep="\t", index=False)
filtered.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/8_gold_disease_feature.tsv", sep="\t", index=False)
