==> Drug_Repositioning/6_new_disease/5_similarity_scores.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import time

gold_disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col=0)

gold_disease.head()

from sklearn.metrics import jaccard_similarity_score
df_jac = pd.DataFrame()
disease_id = gold_disease.columns.tolist()
for i in disease_id:
    desc1 = gold_disease[i]
    for j in disease_id:
        desc2 = gold_disease[j]
        jac_sim = jaccard_similarity_score(desc1,desc2)
        
        temp_df = pd.DataFrame([[i, j, jac_sim]], columns = ['Disease_ID1','Disease_ID2', 'Similarity_Score'])
        df_jac = df_jac.append(temp_df)

df_jac.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_genetic.tsv", sep='\t', index=False)
df_jac.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_genetic.tsv", sep='\t', index=False)

df_jac = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_genetic.tsv")

df_jac[df_jac.Similarity_Score != 1].sort_values('Similarity_Score').head()

df_mim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/MimMiner.txt", header=None)

df_mim.head(2)

omim_id = df_mim[0].tolist() # 첫번 째  컬럼이 omim의 id임.
df_omim2 = df_mim[[x+1 for x in range(5080)]] # 1부터 마지막 column만 있는 dataframe 만들기.
df_omim2.columns = omim_id
df_omim2.index = omim_id

int_disease_id = [int(x) for x in disease_id]

temp = df_omim2.filter(items = int_disease_id)
temp2 = temp.filter(items = int_disease_id, axis=0)

temp2.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_mimminer.tsv", sep='\t')
temp2.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_mimminerc.tsv", sep='\t')
