==> Drug_Repositioning/6_new_disease/6_indep_similarities.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import time

'''2017.4.21. indep disease와 gold disease간의 phenotype, ge similarity 구해야함.'''

gold_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/new_predict/1_gold_learning_asso.tsv")
indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/new_predict/3_indep_learning_asso.tsv")

gold_ge = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/1_disease_disease_ge_sim_mat.tsv", index_col=0)
gold_phe = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/1_disease_disease_phe_sim_mat.tsv", index_col=0)

print len(gold_ge)
print len(gold_phe)
print gold_ge.columns.tolist() == gold_phe.columns.tolist()
gold_ge.head(1)

gold_di_id = gold_asso['disease_id'].drop_duplicates()
print len(gold_di_id)
indep_di_id = indep_asso['disease_id'].drop_duplicates()
print len(indep_di_id)

indep_ge_sim_mat = gold_ge.filter(items = indep_di_id.tolist(), axis = 0)
indep_phe_sim_mat = gold_phe.filter(items = indep_di_id.tolist(), axis = 0)

indep_ge_sim_mat.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/3_indep_di_gold_di_geSim_mat.tsv", sep='\t')
indep_phe_sim_mat.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/3_indep_di_gold_di_pheSim_mat.tsv", sep='\t')

indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/10_PREDICT_association_indep.tsv")

indep_disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col=0)

#gold disease의 similarities. 여기서 indep만 빼기.
df_jac = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_genetic.tsv")
df_mim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_mimminerc.tsv", index_col=0)

print len(indep_asso)
print len(indep_asso['drug'].drop_duplicates())
print len(indep_asso['disease'].drop_duplicates())
indep_asso.head()

indep_disease_id= indep_asso['disease'].drop_duplicates().tolist()

#genetic sim
temp = df_jac[df_jac['Disease_ID1'].isin(indep_disease_id)]
temp = temp[temp['Disease_ID2'].isin(indep_disease_id)]

print len(temp)
print len(temp['Disease_ID1'].drop_duplicates())
print len(temp['Disease_ID2'].drop_duplicates())

temp.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_genetic.tsv", sep='\t', index=False)
temp.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_genetic.tsv", sep='\t', index=False)

#phenotypic sim
str_disease_id = [str(x) for x in indep_disease_id]
temp2 = df_mim.filter(items = str_disease_id)
temp2 = temp2.filter(items = indep_disease_id, axis=0)

temp2.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_mimminerc.tsv", sep='\t')
temp2.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_mimminerc.tsv", sep='\t')

#disease similairities 합치기
disease_ph = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_mimminerc.tsv", index_col=0)
disease_ge = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_genetic.tsv")

disease_ph2 = pd.DataFrame()
for i in disease_ph.columns.values:
    for j in disease_ph.index.values:
        temp_df = pd.DataFrame(data = [[int(i), int(j), disease_ph[[i]].loc[j].values[0]]], columns = ['Disease_ID1', 'Disease_ID2', 'Similarity_Score'] )
        disease_ph2 = disease_ph2.append(temp_df)
df_disease_sim = pd.merge(left = disease_ph2, right= disease_ge, left_on = ['Disease_ID1', 'Disease_ID2'], right_on = ['Disease_ID1', 'Disease_ID2'])
df_disease_sim = df_disease_sim.rename(columns = {'Similarity_Score_x':'PheSim', 'Similarity_Score_y':'GeSim'})
df_disease_sim.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_whole.tsv", sep='\t', index=False)
df_disease_sim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_whole.tsv", sep='\t', index=False)
