==> Drug_Repositioning/6_new_disease/7_umls_id_to_mesh.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime

umls_mesh = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/DigGeNet/map_umls_2_mesh.txt")
umls_mesh.diseaseId = umls_mesh.diseaseId.str.slice(start =5)
umls_mesh.head(1)

gold_di = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/1_gold_disease_id.tsv", header=None)

omim_umls = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/1_omim_umls_mapping_association.tsv")

omim_umls.head(1)

# omim_umls mapping 파일에서, gold disease에 해당하는 umls id를 구한다.
filtered = omim_umls[omim_umls.OMIM_ID.isin(gold_di[0])]
print len(filtered.OMIM_ID.drop_duplicates())
filtered.head(1)

# umls_mesh mapping 파일에서, gold disease에 해당하는 mesh id를 찾는다.
fil2 = umls_mesh[umls_mesh.diseaseId.isin(filtered['UMLS_concept_ID'])]

fil2.head()

print len(set(fil2.diseaseId))
print len(fil2)

merged = pd.merge(left = filtered, right = umls_mesh, left_on = 'UMLS_concept_ID', right_on = 'diseaseId', how='inner')

filtered[(filtered.OMIM_ID == 102400) | (filtered.OMIM_ID == 102500)]

merged[merged.duplicated('UMLS_concept_ID', keep=False)].sort_values('UMLS_concept_ID')

print len(set(filtered[filtered['UMLS_concept_ID'].isin(fil2.diseaseId.str[5:])].OMIM_ID))

ex_oid = filtered[filtered['UMLS_concept_ID'].isin(fil2.diseaseId.str[5:])].OMIM_ID.drop_duplicates()
print len(ex_oid)

filtered[~filtered.OMIM_ID.isin(ex_oid)]

a = ["OMIM:"+str(x) for x in gold_di[0].tolist()]

f = open("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/gold_omim_id.txt", 'w')
for item in a:
    f.write("%s\n" % item)
f.close()
