==> Drug_Repositioning/6_new_disease/8_didiSim(R)_tidy.ipynb <==
# R에서 작업한 didiHpo, didiGoSim을 정리하는 작업.
# 1. header의 순서를 다른 didi sim mat의 순서와 동일하게 한다. (HpoSim같은 경우 이미 해놓음. 이 작업은 didiGO만 한다.)
# 2. gold와 indep 버전을 만든다.

import pandas as pd
import numpy as np

diPhe = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/1_disease_disease_phe_sim_mat.tsv", index_col=0)
diGo = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/diGOSim_resnik_BMA_header.txt", index_col=0)
#diHpo = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/HPO_sim(sqrt, funSimMean,PA)_header.tsv",index_col=0)
#diHpo = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/HPO_sim(sqrt,BMA,PA)_header.txt",index_col=0)

diPhe.head(1)

diGo.head(1)

# R에서 구할 때, NA값이 있는 애들이 있음 --> 0으로 처리
#diHpo.fillna(0.0, inplace=True)
#print diHpo.isnull().any().any()
print diGo.isnull().any().any()
diGo.fillna(0.0, inplace=True)
print diGo.isnull().any().any()

# 이과정은 HPOsim의 경우에만 필요(index, column에 OMIM:이 붙어있었기 때문에)
#diHpo.columns = diHpo.columns.str.slice(start=5,stop=11)
#diHpo.index = [int(x[5:11]) for x in diHpo.index.tolist()]

# diGo의 column순서를 기존의 didi sim mat순서와 맞춘다.
#order = diPhe.columns.values
#diHpo = diHpo[order.tolist()]
order = diPhe.columns.values
diGo = diGo[order.tolist()]

diPhe.columns.tolist() == diGo.columns.tolist()

# diGo의 index 순서를 기존의 didi sim mat순서와 맞춘다.
diGo = diGo.reindex(index = diPhe.index)

diPhe.index.tolist() == diGo.index.tolist()

#diHpo.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/1_disease_disease_HPO_sim_mat(sqrt,BMA,PA).tsv", sep='\t')
#diHpo.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/1_disease_disease_HPO_sim_mat(sqrt,BMA,PA).tsv", sep='\t')
diGo.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/1_disease_disease_GoSim_mat(resnik,BMA).tsv", sep='\t')
diGo.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/1_disease_disease_GoSim_mat(resnik,BMA).tsv", sep='\t')

# 2. ndep 버전을 만든다.

diHpo = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/1_disease_disease_HPO_sim_mat(sqrt,BMA,PA).tsv", index_col=0)
diGo = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/1_disease_disease_GoSim_mat(resnik,BMA).tsv", index_col=0)
indep_pheSim  = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/3_indep_di_gold_di_pheSim_mat.tsv", index_col=0)

print indep_pheSim.shape
indep_order = indep_pheSim.columns.values

indep_hpoSim = diHpo.filter(items = indep_pheSim.index, axis=0)
indep_hpoSim = indep_hpoSim[order.tolist()]
indep_hpoSim = indep_hpoSim.reindex(index = indep_pheSim.index)
print indep_pheSim.columns.tolist() == indep_hpoSim.columns.tolist()
print indep_pheSim.index.tolist() == indep_hpoSim.index.tolist()

indep_goSim = diGo.filter(items = indep_pheSim.index, axis=0)
indep_goSim = indep_goSim[order.tolist()]
indep_goSim = indep_goSim.reindex(index = indep_pheSim.index)
print indep_pheSim.columns.tolist() == indep_goSim.columns.tolist()
print indep_pheSim.index.tolist() == indep_goSim.index.tolist()

indep_hpoSim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/3_indep_di_gold_di_hpoSim_mat.tsv", sep='\t')
indep_goSim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/3_indep_di_gold_di_goSim_mat.tsv", sep='\t')
